Extract fetch of data to another component and clean up a few styles
