+++ src/sass/post-911-gib-status.scss
@@ -1,7 +1,9 @@
 @import "style";
 
 h3 {
-  border-bottom: 1px solid #aeb0b5;
+  border-bottom: 1px solid $color-gray-light;
+  margin: 0 0 .5em 0;
+  padding: 0 0 .25em 0;
 }
 
 h4 {
@@ -10,7 +12,7 @@ h4 {
 
 hr {
   border-style: solid;
-  border-color: #aeb0b5;
+  border-color: $color-gray-light;
   border-width: 1px 0 0 0;
   margin-bottom: 0;
 }
